+++ completion/git-machete.completion.zsh
@@ -79,7 +79,9 @@ _git-machete() {
                     && ret=0
                     ;;
                 (show)
-                    _arguments '1:: :__git_machete_directions_show' && ret=0
+                    _arguments '1:: :__git_machete_directions_show' \
+                        '(-b --branch)'{-b,--branch=}'[Target branch for show]: :__git_machete_list_managed' \
+                    && ret=0
                     ;;
                 (slide-out)
                     _arguments \